initial commit
Lots of rough edges here, and some dead code waiting to be finished and
wired up. The buildsys produces a program called "cow" which implements
an RPC client, exposing the 'msearch', 'search', 'info', and 'multiinfo'
methods. It also implements a 'download' method which queries for
package existence and, if found, downloads and extracts the tarball to
$PWD.